Fixed all the css and html-files after merge with search_bar and filter_flask.
