+++ Gruppe_18/src/main/controller/AccountController.py
@@ -67,16 +67,7 @@ class AccountController:
             flash('You must be logged in to cancel a tour.', 'danger')
             return redirect(url_for('login'))
 
-    def homepage_based_on_usertype(self):
-        tours = self.session.query(Tour).all()
-        if current_user.usertype == "admin":
-            data = self.account_repository.admin_dashboard()
-            print(f"DETT ER DATA: {data}")
-            return render_template('homepage_admin.html', **data)
-        elif current_user.usertype == "guide":
-            return render_template('homepage_guide.html', tours=tours)
-        else:
-            return render_template('homepage.html', tours=tours)
+
 
     def deleting_account(self):
         if current_user.is_authenticated:
