+++ Gruppe_18/src/main/templates/homepage.html
@@ -56,18 +56,16 @@
           padding: 10px 20px;
         }
 
-        /* Left Section with H2 and Search Bar */
         .left-section {
           display: flex;
-          flex-direction: column; /* Stack items vertically */
+          flex-direction: column;
           width: 100vh
         }
 
         .left-section h2 {
-          margin-bottom: 10px; /* Adjust margin as needed */
+          margin-bottom: 10px;
         }
 
-        /* Right Section with Profile Icon */
         .right-section {
           display: flex;
           align-items: center;
@@ -163,7 +161,6 @@
     <header class="header">
         <div class="left-section">
             <h1 class="yourtour">Your Guide</h1>
-            <!-- Legg til funksjonalitet for søkefeltet -->
             <form action="/search" method="get" class="search-bar">
                 <input type="text" name="q" placeholder="Search">
                 <button type="submit">Search</button>
@@ -171,7 +168,6 @@
         </div>
         <button class="filter-button button-placement">filter menu</button>
         <div class="right-section">
-            <!-- Navbar -->
             <nav class="dropdown-menu">
                 <svg xmlns="http://www.w3.org/2000/svg" width="30" height="30" fill="currentColor" class="bi" viewBox="0 0 16 16">
                     <path d="M8 8a3 3 0 1 0 0-6 3 3 0 0 0 0 6Zm2-3a2 2 0 1 1-4 0 2 2 0 0 1 4 0Zm4 8c0 1-1 1-1 1H3s-1 0-1-1 1-4 6-4 6 3 6 4Zm-1-.004c-.001-.246-.154-.986-.832-1.664C11.516 10.68 10.289 10 8 10c-2.29 0-3.516.68-4.168 1.332-.678.678-.83 1.418-.832 1.664h10Z"/>
@@ -195,7 +191,6 @@
     </form>
     <h1>Welcome, choose your adventure</h1>
     <main>
-        <!-- For hver tur i listen av turer, generer HTML-innhold -->
         {% for tour in tours %}
         <div class="tour">
             <h2>{{ tour.title }}</h2>
