+++ Gruppe_18/src/main/templates/homepage_admin.html
@@ -128,11 +128,56 @@
             padding: 2rem;
             background-color: white;
         }
+
+    .dropdown-menu {
+        position: relative;
+    }
+
+    .dropdown-menu:hover .hidden-menu {
+        display: flex;
+        flex-direction: column;
+    }
+
+    .hidden-menu {
+        display: none;
+        width: 20vh;
+        margin-left: -8vh;
+        align-items: center;
+    }
+
+    .dropdown-menu .hidden-menu {
+        position: absolute;
+        width: 10rem;
+        top: 100%;
+        left: 0;
+        z-index: 1;
+        border: 1px solid #ccc;
+        background-color: #fff;
+        border-radius: 5px;
+    }
+    .right-section {
+        display: flex;
+        align-items: center;
+        width: 29vh;
+        height: 15vh;
+        justify-content: center;
+    }
+
     </style>
 </head>
 <body>
     <header class="header">
         <h1 class="yourtour">Your Guide</h1>
+        <div class="right-section">
+            <nav class="dropdown-menu">
+                <svg xmlns="http://www.w3.org/2000/svg" width="30" height="30" fill="currentColor" class="bi" viewBox="0 0 16 16">
+                    <path d="M8 8a3 3 0 1 0 0-6 3 3 0 0 0 0 6Zm2-3a2 2 0 1 1-4 0 2 2 0 0 1 4 0Zm4 8c0 1-1 1-1 1H3s-1 0-1-1 1-4 6-4 6 3 6 4Zm-1-.004c-.001-.246-.154-.986-.832-1.664C11.516 10.68 10.289 10 8 10c-2.29 0-3.516.68-4.168 1.332-.678.678-.83 1.418-.832 1.664h10Z"/>
+                </svg>
+                <div class="hidden-menu">
+                    <a href="/profile">Profile</a>
+                </div>
+            </nav>
+        </div>
     </header>
     <h1>Welcome, to Your Guide administration</h1>
     <main>
