+++ Gruppe_18/src/main/templates/profile.html
@@ -36,7 +36,7 @@
          color: #000;
          text-align: center;
          position: absolute;
-         bottom: 0;
+         margin-top: 2rem;
          left: 0;
          right: 0;
          }
@@ -57,8 +57,8 @@
         margin: 8px 0;
         border: 1px solid #ccc;
         border-radius: 5px;
-        background-color: #f5f5f5; /* Subtle gray background */
-        color: #333; /* Darker text color */
+        background-color: #f5f5f5;
+        color: #333;
     }
 
     .btn {
@@ -78,6 +78,7 @@
     .profile-icon-name-email {
         display: flex;
         flex-flow: column;
+        padding: 1rem;
         }
 
 </style>
